Improve docs and separation
