+++ proxy/api/src/lib.rs
@@ -13,19 +13,15 @@
 )]
 // TODO(xla): Handle all Results properly and never panic outside of main.
 // TODO(xla): Remove exception for or_fun_call lint.
-// TODO(xla): Remove let_underscore_must_use once the issue is resolved: https://github.com/rust-lang/rust-clippy/issues/4980
-// TODO(xla): Remove used_underscore_binding.
 #![allow(
     clippy::clone_on_ref_ptr,
     clippy::expect_used,
     clippy::implicit_return,
     clippy::integer_arithmetic,
-    clippy::let_underscore_must_use,
     clippy::missing_inline_in_public_items,
     clippy::or_fun_call,
     clippy::shadow_reuse,
-    clippy::unseparated_literal_suffix,
-    clippy::used_underscore_binding
+    clippy::unseparated_literal_suffix
 )]
 #![feature(result_flattening)]
