+++ proxy/coco/src/peer.rs
@@ -24,14 +24,18 @@ use radicle_surf::vcs::git;
 use crate::config;
 use crate::project;
 
+/// Peer errors.
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
+    /// Peer accept error.
     #[error(transparent)]
     Accept(#[from] net::peer::AcceptError),
 
+    /// Peer bootstrap error.
     #[error(transparent)]
     Bootstrap(#[from] net::peer::BootstrapError),
 
+    /// Config error.
     #[error(transparent)]
     Config(#[from] config::Error),
 
@@ -39,21 +43,27 @@ pub enum Error {
     #[error("the identity '{0}' already exits")]
     EntityExists(RadUrn),
 
+    /// Entity meta error.
     #[error(transparent)]
     Meta(#[from] entity::Error),
 
+    /// Project creation error.
     #[error(transparent)]
     ProjectCreate(#[from] project::create::Error),
 
+    /// Repo error.
     #[error(transparent)]
     Repo(#[from] repo::Error),
 
+    /// Storage error.
     #[error(transparent)]
     Storage(#[from] storage::Error),
 
+    /// Surf git error.
     #[error(transparent)]
     SurfGit(#[from] git::error::Error),
 
+    /// Verifcation error.
     #[error(transparent)]
     Verification(#[from] entity::HistoryVerificationError),
 }
@@ -171,7 +181,7 @@ impl Api {
             Err(err) => {
                 log::warn!("an error occurred while trying to get 'rad/self': {}", err);
                 None
-            }
+            },
         }
     }
